+++ .gitignore
@@ -147,3 +147,4 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+yarn.lock
