+++ res/prefixes.txt
@@ -1,0 +1,163 @@
+1.0.0.0/24
+1.0.128.0/17
+1.1.1.0/24
+1.1.128.0/17
+1.12.0.0/14
+1.22.0.0/15
+1.38.0.0/15
+1.4.128.0/17
+1.52.0.0/14
+1.66.0.0/15
+103.204.180.0/22
+103.209.140.0/22
+103.21.244.0/22
+103.212.224.0/22
+103.217.92.0/22
+103.219.164.0/22
+103.22.200.0/22
+103.220.16.0/22
+103.229.84.0/22
+103.230.64.0/22
+103.240.136.0/22
+103.243.20.0/22
+103.243.96.0/22
+103.248.80.0/22
+103.249.236.0/22
+103.250.136.0/22
+103.251.216.0/22
+103.253.108.0/22
+103.31.4.0/22
+103.81.228.0/24
+104.153.196.0/22
+104.16.0.0/12
+104.171.160.0/20
+104.192.32.0/21
+104.200.128.0/19
+104.200.96.0/20
+104.201.64.0/18
+104.204.0.0/16
+104.247.32.0/19
+104.37.192.0/21
+104.64.0.0/10
+105.112.0.0/12
+105.128.0.0/11
+105.184.0.0/14
+105.228.0.0/15
+106.75.0.0/16
+106.96.0.0/13
+107.172.0.0/14
+108.162.192.0/18
+109.76.0.0/14
+109.86.0.0/15
+110.174.0.0/15
+110.36.0.0/14
+12.0.0.0/8
+13.104.0.0/14
+139.28.216.0/22
+139.38.0.0/16
+139.61.0.0/16
+14.0.128.0/17
+14.32.0.0/11
+141.101.64.0/18
+141.226.184.0/21
+142.167.0.0/16
+142.229.0.0/16
+160.131.0.0/16
+160.154.0.0/15
+161.213.0.0/16
+161.215.0.0/16
+161.217.0.0/16
+161.239.0.0/16
+161.8.0.0/16
+162.158.0.0/15
+162.208.96.0/22
+162.210.96.0/21
+162.212.24.0/21
+162.213.16.0/22
+162.213.188.0/22
+162.216.44.0/22
+162.217.248.0/22
+162.247.240.0/22
+162.249.240.0/21
+162.251.80.0/22
+162.41.0.0/16
+162.93.0.0/16
+163.171.192.0/18
+164.115.0.0/16
+164.240.0.0/12
+165.166.0.0/16
+165.227.0.0/16
+165.84.128.0/18
+166.21.0.0/16
+166.95.0.0/16
+171.224.0.0/11
+172.111.128.0/17
+172.242.0.0/15
+172.64.0.0/13
+172.83.152.0/21
+172.93.128.0/17
+172.94.0.0/17
+172.96.160.0/23
+173.16.0.0/12
+173.180.0.0/14
+173.195.64.0/20
+173.211.0.0/17
+173.215.0.0/17
+173.222.0.0/15
+173.245.48.0/20
+173.45.0.0/18
+173.47.0.0/16
+174.124.0.0/15
+174.192.0.0/10
+175.100.0.0/17
+175.101.0.0/16
+175.176.0.0/17
+175.180.0.0/14
+176.118.112.0/20
+176.122.128.0/18
+176.123.192.0/20
+176.212.0.0/14
+176.232.0.0/15
+176.28.128.0/17
+176.29.0.0/16
+176.33.0.0/16
+176.47.0.0/16
+176.54.0.0/15
+176.56.0.0/19
+176.57.248.0/21
+176.62.0.0/19
+176.62.176.0/20
+176.65.96.0/19
+185.146.172.0/22
+188.114.96.0/20
+189.240.0.0/12
+190.40.0.0/17
+190.93.240.0/20
+195.242.122.0/23
+196.176.0.0/14
+197.136.0.0/14
+197.160.0.0/13
+197.234.240.0/22
+198.217.250.0/23
+198.41.128.0/17
+198.61.96.0/19
+199.27.128.0/21
+2.176.0.0/12
+2.92.0.0/14
+200.75.160.0/20
+23.227.32.0/19
+5.101.112.0/20
+5.16.0.0/14
+5.252.236.0/22
+5.254.0.0/17
+5.254.160.0/21
+5.53.128.0/17
+5.8.16.0/21
+6.0.0.0/8
+64.68.192.0/20
+65.110.56.0/21
+66.235.200.0/24
+68.67.64.0/20
+8.0.0.0/9
+8.244.0.0/14
+91.234.214.0/24
